bug: Vérification des droits sur le module PMSI dans le check des classes CCodageCCAM et CActeCCAM
git-svn-id: svn://svn.code.sf.net/p/mediboard/code/trunk@27589 a21d6be2-e512-0410-a060-91f5a2946631

+++ modules/dPsalleOp/templates/inc_manage_codes.tpl
@@ -238,7 +238,7 @@
                         {{tr}}Lock{{/tr}}
                       </button>
                     {{/if}}
-                    {{if !$count_actes_by_prat}}
+                    {{if !$count_actes_by_prat && !$_codage->locked}}
                       <button type="button" class="notext trash"
                               onclick="deleteCodages({{$_codage->praticien_id}})">
                         {{tr}}Delete{{/tr}}
